# src/Trips.Api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Trips.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5099/"
    }
  ],
  "paths": {
...
    "/trips/{id}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "operationId": "GetTrip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Trips"
        ],
        "operationId": "DeleteTrip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/trips/{tripId}/optimise": {
      "post": {
        "tags": [
          "Optimisation"
        ],
        "operationId": "Optimise",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimiseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueRunResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/runs/{runId}": {
      "get": {
        "tags": [
          "Optimisation"
        ],
        "operationId": "GetRun",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimisationRunDtoWithSolution"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/trips/{tripId}/lock-solution": {
      "post": {
        "tags": [
          "Advanced"
        ],
        "operationId": "LockSolution",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockSolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/locked-solution": {
      "get": {
        "tags": [
          "Advanced"
        ],
        "operationId": "GetLockedSolution",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolutionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/whatif": {
      "post": {
        "tags": [
          "Advanced"
        ],
        "operationId": "WhatIf",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueRunResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/cost-split": {
      "get": {
        "tags": [
          "Advanced"
        ],
        "operationId": "CostSplit",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostSplitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Advanced"
        ],
        "operationId": "CostSplitWithInputs",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/CostSplitInputsDto"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostSplitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/return-leg": {
      "post": {
        "tags": [
          "Advanced"
        ],
        "operationId": "ReturnLeg",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ReturnLegRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnLegResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CostSplitInputsDto": {
        "required": [
          "fuelPricePerLitre",
          "fuelEconomyLPer100Km",
          "tolls"
        ],
        "type": "object",
        "properties": {
          "fuelPricePerLitre": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "fuelEconomyLPer100Km": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "tolls": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/TollSegmentDto"
            }
          }
        }
      },
      "CostSplitResponse": {
        "required": [
          "tripId",
          "solutionId",
          "entries",
          "totalCost",
          "totalFuel",
          "totalTolls",
          "fuelPricePerLitre",
          "fuelEconomyLPer100Km"
        ],
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "solutionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostSplitEntry"
            }
          },
          "totalCost": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalFuel": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalTolls": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "fuelPricePerLitre": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "fuelEconomyLPer100Km": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "EnqueueRunResponse": {
        "required": [
          "runId"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "LockSolutionRequest": {
        "required": [
          "runId",
          "paretoIndex"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "paretoIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ObjectiveWeightsDto": {
        "required": [
          "driveTime",
          "stopCount",
          "walkAndPt",
          "arrivalSpread",
          "fairness"
        ],
        "type": "object",
        "properties": {
          "driveTime": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "stopCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "walkAndPt": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "arrivalSpread": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "fairness": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "OptimisationRunDtoWithSolution": {
        "required": [
          "run",
          "solution"
        ],
        "type": "object",
        "properties": {
          "run": {
            "$ref": "#/components/schemas/OptimisationRunDto"
          },
          "solution": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SolutionDto"
              }
            ]
          }
        }
      },
...
      "OptimiseRequest": {
        "required": [
          "weights"
        ],
        "type": "object",
        "properties": {
          "weights": {
            "$ref": "#/components/schemas/ObjectiveWeightsDto"
          },
          "solver": {
            "$ref": "#/components/schemas/SolverKind"
          }
        }
      },
...
      "ParticipantWithNodesDto": {
        "required": [
          "id",
          "tripId",
          "displayName",
          "homeLongitude",
          "homeLatitude",
          "hasCar",
          "seats",
          "preferences",
          "candidateNodes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "homeLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "homeLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "hasCar": {
            "type": "boolean"
          },
          "seats": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesDto"
          },
          "candidateNodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateNodeDto"
            }
          }
        }
      },
...
      "ReturnLegRequest": {
        "required": [
          "requests"
        ],
        "type": "object",
        "properties": {
          "requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnRequestDto"
            }
          }
        }
      },
      "ReturnLegResponse": {
        "required": [
          "solutions"
        ],
        "type": "object",
        "properties": {
          "solutions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SolutionDto"
            }
          }
        }
      },
...
      "SolutionDto": {
        "required": [
          "id",
          "optimisationRunId",
          "label",
          "objective",
          "objectiveTerms",
          "routes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "optimisationRunId": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "objective": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "objectiveTerms": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          "routes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DriverRouteDto"
            }
          }
        }
      },
      "SolverKind": {
        "enum": [
          "orTools",
          "heuristic"
        ],
        "default": "orTools"
      },
...
      "TollSegmentDto": {
        "required": [
          "fromStopId",
          "toStopId",
          "amount"
        ],
        "type": "object",
        "properties": {
          "fromStopId": {
            "type": "string",
            "format": "uuid"
          },
          "toStopId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "TripDetailDto": {
        "required": [
          "id",
          "name",
          "destinationName",
          "destinationLongitude",
          "destinationLatitude",
          "departAt",
          "arrivalWindowEarliest",
          "arrivalWindowLatest",
          "ownerId",
          "createdAt",
          "lockedSolutionId",
          "participants"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "destinationName": {
            "type": "string"
          },
          "destinationLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "destinationLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "departAt": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowEarliest": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowLatest": {
            "type": "string",
            "format": "date-time"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lockedSolutionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantWithNodesDto"
            }
          }
        }
      },
      "TripDto": {
        "required": [
          "id",
          "name",
          "destinationName",
          "destinationLongitude",
          "destinationLatitude",
          "departAt",
          "arrivalWindowEarliest",
          "arrivalWindowLatest",
          "ownerId",
          "createdAt",
          "lockedSolutionId",
          "participantCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "destinationName": {
            "type": "string"
          },
          "destinationLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "destinationLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "departAt": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowEarliest": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowLatest": {
            "type": "string",
            "format": "date-time"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lockedSolutionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "participantCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "WhatIfRequest": {
        "required": [
          "dropParticipantIds",
          "addParticipants",
          "newWeights"
        ],
        "type": "object",
        "properties": {
          "dropParticipantIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "addParticipants": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/AddParticipantRequest"
            }
          },
          "newWeights": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ObjectiveWeightsDto"
              }
            ]
          }
        }
      }
    }
  },
...
}